Improve code for FPC cocompilation compatibility

+++ Source/TaurusTLS_BIO.pas
@@ -428,7 +428,7 @@ begin
   if ASize = 0 then Exit;
   CheckCanRead;
 
-  lToRead := IndyMin(ASize, Pending);
+  lToRead := IndyMin(int64(ASize), int64(Pending));
   if lToRead = 0 then Exit;
 
   SetLength(lBuf, AChunkSize);
@@ -436,7 +436,7 @@ begin
   while lToRead > 0 do
   begin
     // Reads from the BIO
-    lReadSize := Read(lBuf[0], IndyMin(lToRead, AChunkSize));  //PALOFF "NativeUInt cast to Integer"
+    lReadSize := Read(lBuf[0], IndyMin(int64(lToRead), int64(AChunkSize)));  //PALOFF "NativeUInt cast to Integer"
     if lReadSize = 0 then Break;
 
     AStream.WriteBuffer(lBuf[0], Integer(lReadSize));
